requirements-llm.txt
Flask==3.0.3
flask-cors==4.0.1
openai==1.77.0
PyMySQL==1.1.1
PyYAML==6.0.2
